app/requirements.txt
locustio==0.7.5


